requirements.txt
﻿fastapi
numpy
pandas
uvicorn
gunicorn
matplotlib
ydata-profiling
scikit-learn
apache-airflow
dvc
great_expectations
